Move validation code from utils to validation directory
